+++ MetaStalk/__init__.py
@@ -1,6 +1,6 @@
-"""MetaStalk
+"""MetaStalk.
 ---
 MetaStalk is a program that parse image metadata and creates charts from it.
 """
-__version__ = "v2.0.0"
+__version__ = "v2.1.0"
 __author__ = "Cyb3r Jak3"