+++ MetaStalk/modules/GPSCheck.py
@@ -1,11 +1,11 @@
-"""Makes geo chart with plots of gps data"""
+"""Makes geo chart with plots of GPS data"""
 import logging
 import plotly.express as px
 
 log = logging.getLogger("MetaStalk")
 
 
-def GPS_Check(photos: list) -> px.scatter_mapbox:
+def gps_check(photos: list) -> px.scatter_mapbox:
     """GPS_Check
 
     Takes a list of photos and creates a geo plot of them
@@ -13,7 +13,7 @@ def GPS_Check(photos: list) -> px.scatter_mapbox:
     Arguments:
         photos {list} -- A list of dictionaries with phot information.
 
-    Returns:
+    Returns
         px.scatter_mapbox -- Map plot with photos plotted.
     """
     log.info("Starting GPS Chart")
@@ -26,14 +26,16 @@ def GPS_Check(photos: list) -> px.scatter_mapbox:
             gps_photos.append(each["item"])
             lats.append(float(each["Latitude"]))
             longs.append(float(each["Longitude"]))
-            log.debug("%s has gps data", each["item"])
+            log.debug("%s has GPS data", each["item"])
+        else:
+            log.info("%s has no GPS data", each["item"])
 
     points = []
     for x, _ in enumerate(gps_photos):
         points.append((lats[x], longs[x]))
 
     fig = px.scatter_mapbox(lon=longs, lat=lats, hover_name=gps_photos,
-                            title="Geo Locations")
-    fig.update_layout(mapbox_style="open-street-map")
+                            title="Geo Locations",)
+    fig.update_layout(mapbox_style="open-street-map", title_x=0.5)
 
     return fig